# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Actor:
    properties:
      client_id:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  response.Meta:
    properties:
      actor:
        $ref: '#/definitions/response.Actor'
      request_id:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.ResponseError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ResponseError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
...
  users.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  users.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/users.UserResponse'
    type: object
  users.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  users.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    required:
    - first_name
    - last_name
    type: object
  users.UserResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      phone_verified:
        type: boolean
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@gogin.com
    name: API Support
  description: |-
    A comprehensive REST API built with Go and Gin framework
    This API provides user authentication, profile management, and admin operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gogin API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get a paginated list of all users (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    page:
                      type: integer
                    total:
                      type: integer
                    total_pages:
                      type: integer
                    users:
                      items:
                        $ref: '#/definitions/users.UserResponse'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Admin
  /users/{id}:
    delete:
      description: Delete a user account (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Admin
    get:
      description: Get a specific user by their ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    user:
                      $ref: '#/definitions/users.UserResponse'
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update a user's profile information (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    user:
                      $ref: '#/definitions/users.UserResponse'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Admin
  /users/{id}/status:
    put:
      consumes:
      - application/json
      description: Update a user's status (active, inactive, or suspended) (admin
        only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Status update details (status must be: active, inactive, or
          suspended)'
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    status:
                      type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - Admin
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive access and refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/users.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      summary: User login
      tags:
      - Users
  /users/logout:
    post:
      description: Logout the authenticated user and invalidate their session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Users
  /users/me:
    delete:
      description: Delete the authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - Users
    get:
      description: Get the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    user:
                      $ref: '#/definitions/users.UserResponse'
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: Profile update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    user:
                      $ref: '#/definitions/users.UserResponse'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Change the authenticated user's password
      parameters:
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    user:
                      $ref: '#/definitions/users.UserResponse'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      summary: Register a new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...